Inline flashoffer demo copy
