+++ app/flashoffer-demo/src/sections/OverviewSection.tsx
@@ -1,7 +1,6 @@
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
 import { Section } from "flashoffer-react";
-import { OVERVIEW_PARAGRAPHS } from "./content";
 
 export interface OverviewSectionProps {
   overviewMeta: string;
@@ -20,17 +19,26 @@ export function OverviewSection({ overviewMeta }: OverviewSectionProps) {
         title="Narrate launches with reusable sections"
       >
         <Typography color="text.secondary">
-          Combine headlines and supporting copy without rebuilding layouts from
-          scratch.
+          Combine headlines and supporting copy while staying independent from
+          opaque review cycles.
         </Typography>
-        {OVERVIEW_PARAGRAPHS.map((paragraph) => (
-          <Typography key={paragraph} color="text.secondary">
-            {paragraph}
-          </Typography>
-        ))}
+        <Typography color="text.secondary">
+          Flashoffer-react emerged after repeated takedowns of fine art on
+          mainstream networks. We respect law and order, yet opaque filters
+          squeeze the livelihoods of artists.
+        </Typography>
+        <Typography color="text.secondary">
+          We understand the tension between shielding minors from explicit
+          material and giving artists room to breathe. The framework offers
+          reusable, policy-free layouts so you can advertise art on your own
+          terms.
+        </Typography>
+        <Typography color="text.secondary">
+          Compose campaigns and deploy them to any Docker-ready cloud in
+          minutes. While any tool can be misused, we trust creators to stand by
+          one another more often than not.
+        </Typography>
       </Section>
     </Box>
   );
 }
-
-export { OVERVIEW_PARAGRAPHS };
